support/dockerfiles: Provide generic build envs for NGINX and Redis
This commit introduces a shared base docker image which is used for
NGINX and Redis build environments for FlexOS.  These dockerfiles are
intended to create a reproducible build environment for said
applications.  These will be primarily used for Figure 6. where many
image permutations will be built and so a reproducible environment
is necessary.

Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>