Rename experiment structure
This commit renames each experiment's canonical name from:

  fig_XX_short-description

to

  fig-XX_short-description

The intent is to use the underscore as a separation between the
identifier and the short description.  This will make it easy to
split since we can use the underscore.  Later commits will introduce
Makefile targets which rely on splitting the name this way so we can
achieve entrypoints such as:

  make fig-XX

instead of having to use the full canonical name.

Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>